refactor: moved functions in /utils to /helpers
