+++ src/helpers/verifyStoredToken.js
